fix(android): autolinking working for expo
